Don't replace the GestureDetector
Before this change we used to replace the GestureDetector when the keyboard
layout changed.

That broke the long press up action though.

Now we only replace the GestureListener when the keyboard changes, and long
pressing and releasing over some number actually works!

Next step is to test this on a real device, update the README and merge to
master!