Add popup keyboards for e, E and A
Also, this change enables us to have functioning popup keyboards for keys on the
right hand side of the keyboard as well. Tested and found working.

The result of that test is only documented here; I couldn't come up with an
actual use case for it so there are no keys utilizing that functionality right
now.