Set max height relative to screen
Before this change we set the max height relative to how much space we got in
the layout process.

However, since we're laid out multiple times, and the constraints we get depend
on what we returned during the last run, doing it that way resulted in us using
only a quarter or so of the screen height and leaving half the screen empty in
portrait mode.